+++ controllers/perses/service_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package perses
 
 import (
 	"context"
@@ -25,93 +25,90 @@ import (
 	common "github.com/perses/perses-operator/internal/perses/common"
 	subreconciler "github.com/perses/perses-operator/internal/subreconciler"
 	logger "github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
-var cmlog = logger.WithField("module", "configmap_controller")
+var slog = logger.WithField("module", "service_controller")
 
-func (r *PersesReconciler) reconcileConfigMap(ctx context.Context, req ctrl.Request) (*ctrl.Result, error) {
+func (r *PersesReconciler) reconcileService(ctx context.Context, req ctrl.Request) (*ctrl.Result, error) {
 	perses := &v1alpha1.Perses{}
 
 	if r, err := r.getLatestPerses(ctx, req, perses); subreconciler.ShouldHaltOrRequeue(r, err) {
 		return r, err
 	}
 
-	configName := common.GetConfigName(perses.Name)
-
-	found := &corev1.ConfigMap{}
-	err := r.Get(ctx, types.NamespacedName{Name: configName, Namespace: perses.Namespace}, found)
+	found := &corev1.Service{}
+	err := r.Get(ctx, types.NamespacedName{Name: perses.Name, Namespace: perses.Namespace}, found)
 	if err != nil && apierrors.IsNotFound(err) {
-		cm, err := createPersesConfigMap(r, perses)
+
+		ser, err := r.createPersesService(perses)
 		if err != nil {
-			cmlog.WithError(err).Error("Failed to define new ConfigMap resource for perses")
+			slog.WithError(err).Error("Failed to define new Service resource for perses")
 
 			meta.SetStatusCondition(&perses.Status.Conditions, metav1.Condition{Type: common.TypeAvailablePerses,
 				Status: metav1.ConditionFalse, Reason: "Reconciling",
-				Message: fmt.Sprintf("Failed to create ConfigMap for the custom resource (%s): (%s)", perses.Name, err)})
+				Message: fmt.Sprintf("Failed to create Service for the custom resource (%s): (%s)", perses.Name, err)})
 
 			if err := r.Status().Update(ctx, perses); err != nil {
-				cmlog.WithError(err).Error("Failed to update perses status")
+				slog.Error(err, "Failed to update perses status")
 				return subreconciler.RequeueWithError(err)
 			}
 
 			return subreconciler.RequeueWithError(err)
 		}
 
-		cmlog.Infof("Creating a new ConfigMap: ConfigMap.Namespace %s ConfigMap.Name %s", cm.Namespace, cm.Name)
-		if err = r.Create(ctx, cm); err != nil {
-			cmlog.WithError(err).Errorf("Failed to create new ConfigMap: ConfigMap.Namespace %s ConfigMap.Name %s", cm.Namespace, cm.Name)
+		slog.Infof("Creating a new Service: Service.Namespace %s Service.Name %s", ser.Namespace, ser.Name)
+		if err = r.Create(ctx, ser); err != nil {
+			slog.WithError(err).Errorf("Failed to create new Service: Service.Namespace %s Service.Name %s", ser.Namespace, ser.Name)
 			return subreconciler.RequeueWithError(err)
 		}
 
 		return subreconciler.RequeueWithDelay(time.Minute)
-	}
+	} else if err != nil {
+		slog.WithError(err).Error("Failed to get Service")
 
-	if err != nil {
-		cmlog.WithError(err).Error("Failed to get ConfigMap")
 		return subreconciler.RequeueWithError(err)
 	}
 
 	return subreconciler.ContinueReconciling()
 }
 
-func createPersesConfigMap(r *PersesReconciler, perses *v1alpha1.Perses) (*corev1.ConfigMap, error) {
-	configName := common.GetConfigName(perses.Name)
-	ls, err := common.LabelsForPerses(r.Config.PersesImage, configName, perses.Name)
+func (r *PersesReconciler) createPersesService(
+	perses *v1alpha1.Perses) (*corev1.Service, error) {
+	ls, err := common.LabelsForPerses(r.Config.PersesImage, perses.Name, perses.Name)
 
 	if err != nil {
 		return nil, err
 	}
 
-	persesConfig, err := yaml.Marshal(perses.Spec.Config.Config)
-
-	if err != nil {
-		cmlog.WithError(err).Errorf("Failed to marshal configmap data: ConfigMap.Namespace %s ConfigMap.Name %s", perses.Namespace, configName)
-		return nil, err
-	}
-
-	configData := map[string]string{
-		"config.yaml": string(persesConfig),
-	}
-
-	cm := &corev1.ConfigMap{
+	ser := &corev1.Service{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      configName,
+			Name:      perses.Name,
 			Namespace: perses.Namespace,
 			Labels:    ls,
 		},
-		Data: configData,
+		Spec: corev1.ServiceSpec{
+			Type: corev1.ServiceTypeClusterIP,
+			Ports: []corev1.ServicePort{{
+				Name:       "http",
+				Port:       8080,
+				Protocol:   corev1.ProtocolTCP,
+				TargetPort: intstr.FromInt(8080),
+			}},
+			Selector: ls,
+		},
 	}
 
-	// Set Perses instance as the owner and controller
-	if err := ctrl.SetControllerReference(perses, cm, r.Scheme); err != nil {
+	// Set the ownerRef for the Service
+	// More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/
+	if err := ctrl.SetControllerReference(perses, ser, r.Scheme); err != nil {
 		return nil, err
 	}
-	return cm, nil
+	return ser, nil
 }
